Added life 1Up at a certain score which is shown in the status screen. Some Messages are loaded for Episode 4

+++ src/engine/galaxy/ItemsGalaxy.h
@@ -18,7 +18,7 @@ struct stItemGalaxy
 	unsigned char &m_difficulty;
 	signed char m_lifes;
 	unsigned long  m_points;
-	int m_lifeAt;
+	unsigned long m_lifeAt;
 	int m_drops;
 	unsigned int m_bullets;
 	struct {
